Replace 'this.id' with e.currentTarget.id
I previously used this.id because e.target.id doesn't return
the div.weapon id when user clicks on the image or the label.
this.id works, but I didn't fully understand why.

Adding '{ capture: true }' in the event listener is the key as the
event propagates from the parent down to its decendants instead
of the default (bubbling from children to parent), so the event
gets fired up for the parent first (i.e. div.weapon), which
suits the intention for this case. Then it's just a matter of
suppressing propagation further.